Policy xml files: Change angle: grads->grad
The plural "grads" does not appear to be an acceptable unit
for a CSS `<angle>` type.

"Angle" in the policy files appears in the context of either
CSS Aural `azimuth` or `elevation` properties.

The W3 appendix on Aural stylesheets [1] lists `deg`, `grad` and `rad`
as valid units for `<angle>` types.

[1] https://www.w3.org/TR/CSS21/aural.html#value-def-angle