Back up to develop a 1.6.6 release candidate (Max Java 7). We identified a maintained version of org.cyberneko.html at: net.sourceforge.htmlunit.cyberneko. Upgrading to this version should help us eliminate a bunch of bugs that were present in the OLD unmaintained cyberneko.
This commit swaps these two libraries and sets net.sourceforge.htmlunit.cyberneko
to version 2.24, which is the last version that supports Java 7. Hopefully that
will be good enough for what we want to do with the 1.6.6 release and when we
release 1.7.0 soon (which will require Java 8), we can upgrade to the latest
version of net.sourceforge.htmlunit.cyberneko, which is currently at 2.60.0.